+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+## 0.9.0-alpha.7 (2022-12-01)
+
 **新增**
 
-- 完善更多 `f-menu` 组件功能
 - 新增 `f-breadcrumb-item` 组件 `to` 配置项
+- 完善更多 `f-menu` 组件功能
 
 **优化**